Created new package to represent Tasks
