+++ src/main/java/duke/command/DoneCommand.java
@@ -1,6 +1,10 @@
 package duke.command;
 
-import duke.*;
+import duke.DukeException;
+import duke.Storage;
+import duke.Ui;
+import duke.task.Task;
+import duke.task.TaskList;
 
 public class DoneCommand extends Command {
     protected int taskNumber;
