+++ src/main/java/duke/command/EmptyInputCommand.java
@@ -1,8 +1,8 @@
 package duke.command;
 
 import duke.Storage;
-import duke.TaskList;
 import duke.Ui;
+import duke.task.TaskList;
 
 public class EmptyInputCommand extends Command {
