+++ src/main/java/duke/command/ExitCommand.java
@@ -1,8 +1,8 @@
 package duke.command;
 
 import duke.Storage;
-import duke.TaskList;
 import duke.Ui;
+import duke.task.TaskList;
 
 public class ExitCommand extends Command {
     public ExitCommand() {
