+++ src/main/java/duke/task/TaskList.java
@@ -1,4 +1,4 @@
-package duke;
+package duke.task;
 
 import java.util.ArrayList;
 import java.util.Iterator;
