+++ src/test/java/duke/UiTest.java
@@ -1,5 +1,6 @@
 package duke;
 
+import duke.task.*;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
