+++ text-ui-test/sources.txt
@@ -1,6 +1,6 @@
-C:\Users\jerem\Desktop\NUS_CS\Y2_Sem2\CS2103T\duke\src\main\java\duke.Deadline.java
+C:\Users\jerem\Desktop\NUS_CS\Y2_Sem2\CS2103T\duke\src\main\java\duke.task.Deadline.java
 C:\Users\jerem\Desktop\NUS_CS\Y2_Sem2\CS2103T\duke\src\main\java\duke.Duke.java
-C:\Users\jerem\Desktop\NUS_CS\Y2_Sem2\CS2103T\duke\src\main\java\duke.Event.java
-C:\Users\jerem\Desktop\NUS_CS\Y2_Sem2\CS2103T\duke\src\main\java\duke.Task.java
+C:\Users\jerem\Desktop\NUS_CS\Y2_Sem2\CS2103T\duke\src\main\java\duke.task.Event.java
+C:\Users\jerem\Desktop\NUS_CS\Y2_Sem2\CS2103T\duke\src\main\java\duke.task.Task.java
 C:\Users\jerem\Desktop\NUS_CS\Y2_Sem2\CS2103T\duke\src\main\java\duke.CommandType.java
-C:\Users\jerem\Desktop\NUS_CS\Y2_Sem2\CS2103T\duke\src\main\java\duke.Todo.java
+C:\Users\jerem\Desktop\NUS_CS\Y2_Sem2\CS2103T\duke\src\main\java\duke.task.Todo.java
